* Abstraction for the sql code created. (common/sql) * Login sql server using the new code. (testing is needed) * Cleaned up some code and warnings.
git-svn-id: http://eathena.googlecode.com/svn/branches/tmpsql@10753 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/common/ers.c
@@ -40,9 +40,10 @@
 \*****************************************************************************/
 #include <stdlib.h>
 
-#include "ers.h"
+#include "../common/cbasetypes.h"
 #include "../common/malloc.h" // CREATE, RECREATE, aMalloc, aFree
 #include "../common/showmsg.h" // ShowMessage, ShowError, ShowFatalError, CL_BOLD, CL_NORMAL
+#include "ers.h"
 
 #ifndef DISABLE_ERS
 /*****************************************************************************\
@@ -244,7 +245,7 @@ static void ers_obj_free_entry(ERS self, void *entry)
  * @see ERS_impl#size
  * @see ERS_impl::vtable#entry_size
  */
-static uint32 ers_obj_entry_size(ERS self)
+static size_t ers_obj_entry_size(ERS self)
 {
 	ERS_impl obj = (ERS_impl)self;
